GOT A TRIGGER WORKING!!!
No chatgpt was used in the making of this trigger. Pure documentation
and mental fortitude. It was amazing, simple but amazing. Oh, and I also
got a bunch of storedProcedures up and running. They had a very similar
syntax to those from T-Sql, but not exactly.

Also removed all 'IF NOT EXISTS' statements from the tables and stored procedures as they were made redundant in mysql v8.0.0 onwards.

Anyway, day made!!!!