ADDED THE VIEWS!
which wasn't quite as hard as i thought as they only virtual tables. The
functions proved abit difficult and thus were put off for a while. Was
able to bundle the entire db-schema into a script and execute it
seamlessly at one which was quite nice.

Looking forward to begining work on the backend implemenattion tommorow.