Created the first endpoint!!(technically 2nd)
Created the register user endpoint in the rather simple way with no
validation or authorization.

Had to rewrite the database schema, in that the initial INT data type
associated with the id column had to truncate the long uuid generated
id. Converted this to VARCHAR(255) to house it without truncating it.

build the routes and collections to help in eventual modularization of
the application.