did the CR operations on the database
next phase will involve working on the update and deacticate(delete)

some bugs were identified manually in the loginUser controller, whereby
if i were to input a username/email that does not exist, an error will
be caught instead of returnng that that user does not exist. Will look
on how to handle this while testing the application later on.

Ofcourse the fragile premise being that i able to remember that ;(