Fixed the BUGGG!!!
10 years later, I was able to fix the bug. it resulted due to two main
issues:

 1.)Wrong versioning - I was previously using v5.0.1 for both express
and @types/express, which are core packages in the app. I downgraded
both of these, since these are the latest packages currently and might
have some unseen breaking changes. Now using express@4.19.2 and
@types/express@4.17.21. This removed the function overload error that
was in the users.routes.ts

2.)Joi validation error -> I was incorrectly validating a Joi object
with .message while the only requirement was .string() and .required().
Had to change this to messages({'any.required':'blah blah'}) to resolve
that.

3.)Also change db.sql to match code refactor

Premium tias guy, but we did it. Lets gooooo!!!