RBAC(Role Based Access Controll)
Implemented logic for role based access by creating a role column in
users table. This had two possible enumerations types, admin and user,
which was enforced both in mysql and in .ts files.

RBAC was used to prevent users who are not admins from accessing two
routes, namely:
  - getUsersByEmail
  - getUsers

This led to change of code within app, moreso in the models and routes.

Also updated the README to reflect the changes. Turns out casbin was
overkill!!!!