Hash user password before storage in db
Began to hash user password before storing them in the database. This
will help tim improve security for users.

Could either hash this on the database side using
ENCRYTP_BY_PASSPHRASE() or just do in the backend sid. I opted to do  it
this way as it would save me alot of uncessary lines. It also makes it
easier to decrypt and compare them when the user is loggin in.