Change reactivating acc logic
Made it so that there is no endpoint for users to reactivate their acc,
Only just one to login. If the user tries to login and they had
deactivated their acc, they will be reactivated, else if it is just a
normal user logging in, all will flow normally.

This also involved making some changes to the user interface to store
the isDeactivated value.