Allow only admins to delete user accounts
Incorporate functionality for admins to delete user accounts, This will
be either using the users email or username which are unique for each
user. If the user is not an admin, appropriate error messages will be
returned.

While doing this part, realises there are massive chunks of code that
are quite redundant, say the part for checking whether an input is an
email or username. Irregardless of either, the steps followed for each
are massively similar, look into alternative ways to make this shorter
somehow to reduce copy pasting.