Add token authorization
Added token authorization for two endpoints:
- getUsers
- deleteUser

Information passed within this token will be used to track whether role
is an admin, and if so continue with the operation, else the operation
will be aborted.

Came up with a question will building this module, why when i used
positive bias within the function, it will not work as expected, while
when I use negative bias everything works perfectly. This is what I
mean:

This works fine:
```typescript
        // verify token is valid
        if(!token){
            // return error if no token
            return response.status(401).json({error:`Access forbidden. Please contact your administrator!`})

        }
        const decodedData = jwt.verify(token,process.env.SECRET as string) as UserPayload

```

This breaks and does not run:
```typescript
    if(token){
    const decodedData = jwt.verify(token,process.env.SECRET as string)
    as UserPayload
    }

    return response.status(401).json({errro:`Access forbidden`})

```

Look into that when possible.