Setup basic UI with PrimeNg
Didnt want to invest too much time in designing the system from scratch,
so I decided to go with vanilla CSS and HTML, but i proved to be quite
daunting when it came to the signup page as it required a stepper.

Tailwind CSS is too bloated and I dont want 2000LOC in my files. After
various reddit threads, I cam upon Prime Ng. Seems pretty decent and
handy to set up, Im giving it a try.

Also realized it has alterbative for other frameworks, i.e React and
Vue, and will prove worthy if I were to learn those later on.