Add documentation into each directory
Based on teams-for-linux documenatation approach, where they place a
small .md file in each directry of the application, and it explains some
key concepts of how the code is position.

This has helped to drastically reduce the number of comments in my code
files. Unsure if this is the best way for documenation of code at the
moment, but i'll stick with it until I figured out a way to do better.