Begin working on background sevices
When a new user registers into the stem, i Want then to get an email
that asks them to verify their email and reflect that to the db. This
will utilize nodemailer and handlebars for conducting the process in the
background. I was able to get it runnign, but manually, next is
automating the whole process, when a user registeres.

I am yet to write any tests, but hey!!