feat(src): implement logic for welcome email
integrate logic and data flow to ensure newly signed in users are sent a
welcome email for joining the email. Created helpers directory that will
be used to implement logic for all the other email services.

TODO: changed logic to use stored procedure and not queries in the
welcome-email.service
WIN: began using conventional commit messages, they are pretty nice and
descriptive. Also did this using handlebars and not ejs as I am used to.
Preferred them as they are closer tho the Jinja2 syntax that I am highly
accustomed to