feat(database):refined the db setup script
made some changes to the script in charge of creating or removing the
database. set it up manually and i think it works pretty nicely,

WINS: ayoo. scripting takes some time before making it work, but after
that its literally effortless. Can easily setup and teardown the db in
seconds