feat(helpers): add the validation helper
removed the repetitive user input validation section from my
controllers, and placed it in a useful helpers/ file. this allows less
LOC per files as i call it within just a single line.

WIN: thought of was of making my users.controllers.ts file smaller and
discovered this is a common practise! Felt nice dicovering that