feat(loggin): added logging to my app
utilizing winston logger, the app is supposed to have the added
functionality of logging all occurrences ongoing in app, requsts
and such to set files, .log and .json files. The log is for
manual lookup and the JSON might be essential if integrated with
services such as elastic search and kibana.

Will look into those later on... a great setback currently is that the
logs do not work*** They appear on the console, but that is all. When
called from the node environement, it also works, but in the controller,
they are not registered in the respective log files.

WINS: first time working with winston, ever! not such a bad tool, the
docs helped a tonne