feat(logging): got the logs working!!
while attending the angular dev fest, i thought of a hack to amke things
work and it worked. Removed logging of registration/login attempts
and only logged it if succesful registered from the db or unsuccesful in
teh validation helper. The error which previously broke the
request/response cycle was eradicated and everything now worked
effortlessly.

WIN: debugging skills always on top!!